remove unused card
